AST to Platelet Ratio (APRI)

A vectorised function to calculate the APRI score, a predictor of hepatic fibrosis.

apri(ast, plt, ast_uln = 40)

Arguments

ast

numeric vector of serum AST levels in IU/l

plt

numeric vector of platelet counts (10^9/l)

ast_uln

single number value for lab upper limit of normal for AST levels (default is 40)

Value

numeric vector of APRI scores

Details

Reference: Lin ZH, Xin YN, Dong QJ, et al. Performance of the aspartate aminotransferase-to-platelet ratio index for the staging of hepatitis C-related fibrosis: an updated meta-analysis. Hepatology 2011; 53:726-736.

Examples

apri(ast = 38, plt = 150, ast_uln = 40)
#> [1] 0.6333333

# if the lab upper limit of normal is 40, ast_uln can be omitted
apri(ast = 160, plt = 75)
#> [1] 5.333333