Pancreas donor risk index

A vectorised function to calculate the Pancreas Donor Risk Index as published by Axelrod et al. By default, the serum creatinine is used in mcmmol/l but this can be changed to mg/dl by setting the optional units parameter to "US"

pdri(age, sex, creat, eth, bmi, height, cva, cit, dcd, intent = c("SPK",
  "PAK", "Other"), units = "SI")

Arguments

age

numeric vector of ages in years

sex

character vector of patient sex ("F" for female, "M" for male)

creat

numeric vector of serum creatinine (µmol/l by default)

eth

character vector of ethnicity, one of "asian", "black" or "other" in each case

bmi

numeric vector of body mass index (BMI)

height

numeric vector of heights in centimetres

cva

numeric vector of whether CVA is cause of death (1 = yes, 0 = no)

cit

numeric vector of cold ischaemic times in hours

dcd

numeric vector of whether donor after circulatory death (1 = DCD, 0 = DBD)

intent

character vector of implant intent, "PAK" for pancreas after kidney or "other"

units

single character string for creatinine units: one of "SI" (for µmol/l) or "US" (for mg/dl)

Value

numeric vector of pancreas donor risk index values

Details

Reference: Axelrod DA, Sung RS, Meyer KH, et al. Systematic evaluation of pancreas allograft quality, outcomes and geographic variation in utilization. Am J Transplant 2010; 10(4):837-45.

Examples

pdri(age = 28, sex = "M", creat = 1.0, eth = "other", bmi = 24,
height = 173, cva = 0, cit = 12, dcd = 0, units = "US") # 1.00
#> [1] 1

pdri(age = 45, sex = "M", creat = 88.4, eth = "other", bmi = 24,
height = 173, cva = 0, cit = 12, dcd = 0, units = "SI") # 1.56
#> [1] 1.559762