PELD score (US units)

A wrapper for peld(), a vectorised function to generate a PELD score for paediatric liver transplant candidates, using mg/dl as the unit for the serum bilirubin level.

peld_US(INR, bili, albumin, listing_age, growth_failure)

Arguments

INR

numeric vector of INR

bili

numeric vector of serum biliruin (mg/dl)

albumin

numeric vector of serum albumin (g/dl)

listing_age

numeric vector of age at the time of listing (years)

growth_failure

numeric vector of whether there is growth failure (1 = yes, 0 = no)

Value

numeric vector of PELD scores

Details

Reference: McDiarmid SV, Anand R, Lindblad AS, et. al. Development of a pediatric end-stage liver disease score to predict poor outcome in children awaiting liver transplantation. Transplantation 2002; 74(2):173-81.

Examples

peld_US(INR = 2, bili = 3.1, albumin = 2.5, listing_age = 2, growth_failure = 1)
#> [1] 18.67756