SOFT score (Survival Outcomes Following Liver Transplantation) (US units)

A wrapper function using US units for the soft() vectorised function to calculate SOFT Scores for predicting patient survival after liver transplantation. The units for donor serum creatinine and recipient serum albumin in g/l.

soft_US(Age, BMI, PrevTx, AbdoSurg, Albumin, Dx, ICU, Admitted, MELD,
  LifeSupport, Encephalopathy, PVThrombosis, Ascites, PortalBleed,
  DonorAge, DonorCVA, DonorSCr, National, CIT)

Arguments

Age

numeric vector of patient ages in years

BMI

numeric vector of patient BMI in kg/m2

PrevTx

numeric vector of number of previous transplants

AbdoSurg

numeric vector of whether previous abdominal surgery (1 = "yes", 0 = "no")

Albumin

numeric vector of serum albumin in g/dl

Dx

numeric vector of whether on dialysis before transplant (1 = "yes", 0 = "no")

ICU

numeric vector of whether patients in intensive care unit before transplant (1 = "yes", 0 = "no")

Admitted

numeric vector of whether admitted to hospital pre-transplant (1 = "yes", 0 = "no")

MELD

numeric vector of MELD scores

LifeSupport

numeric vector of whether on life support pre-transplant (1 = "yes", 0 = "no")

Encephalopathy

numeric vector of whether encephalopathy present (1 = "yes", 0 = "no")

PVThrombosis

numeric vector of whether portal vein thrombosis (1 = "yes", 0 = "no")

Ascites

numeric vector of whether ascites pre-transplant (1 = "yes", 0 = "no")

PortalBleed

numeric vector of whether portal bleeding in 48 hours pre-transplant (1 = "yes", 0 = "no")

DonorAge

numeric vector of donor ages in years

DonorCVA

numeric vector of whether donor cause of death is CVA/stroke (1 = "yes", 0 = "no")

DonorSCr

numeric vector of donor terminal serum creatinine in mg/dl

National

numeric vector of whether national allocation (1 = "yes", 0 = "no")

CIT

numeric vector of cold ischaemic time in hours

Value

numeric vector of SOFT Scores

Details

Reference: Rana A, Hardy MA, Halazun KJ, et al. Survival Outcomes Following Liver Transplantation (SOFT) Score: A Novel Method to Predict Patient Survival Following Liver Transplantation. American Journal of Transplantation 2008; 8:2537-2546.

Examples

soft_US(Age = 35, BMI = 20, PrevTx = 0, AbdoSurg = 1, Albumin = 3.0, Dx = 0,
    ICU = 0, Admitted = 0, MELD = 29, LifeSupport = 0, Encephalopathy = 1,
    PVThrombosis = 0, Ascites = 1, PortalBleed = 0, DonorAge = 44, DonorCVA = 0,
    DonorSCr = 1.2, National = 0, CIT = 8) # 7
#> [1] 7