UK Kidney Donor Risk Index (NHSBT, 2019 version)

A vectorised function to calculate the UK Kidney Donor Risk Index as used in the new national kidney matching scheme implemented in September 2019.

ukkdri(age, height, htn, sex, cmv, gfr, hdays)

Arguments

age

numeric vector of donor age in years

height

numeric vector of donor height in cm

htn

numeric vector of whether donor history of hypertension (1 = yes, 0 = no)

sex

character vector of donor sex ("F" = female, "M" = male)

cmv

numeric vector of whether donor CMV IgG positive (1 = yes, 0 = no)

gfr

numeric vector of donor eGFR at time of donation

hdays

numeric vector of number of days donor in hospital before donation

Value

numeric vector of UK Kidney Donor Risk Index values (2019 version)

Details

The UK KDRI is documented in the UK kidney matching policy which can be found on the NHS Blood & Transplant ODT website at www.odt.nhs.uk

Examples

ukkdri(age = 50, height = 170, htn = 1, sex = "F", cmv = 0, gfr = 90, hdays = 2)
#> [1] 0.9950125